MFTune-db/requirements.txt
HEBO~=0.3.6
matplotlib==3.8.0
numpy==1.24.4
openbox==0.8.4
pandas==2.2.3
psycopg2_binary==2.9.9
pyDOE==0.3.8
pyDOE==0.3.8
rpy2==3.5.17
scikit_learn==1.5.1
scipy==1.10.1
seaborn==0.13.2
statsmodels==0.14.4
tqdm==4.66.4
mysql-connector-python==8.1.0
lingam==1.10.0
causal-learn==0.1.4.3